scaffolded app, created containers and moved files around
