+++ bookr/src/components/Login/Login.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Button from "@material-ui/core/Button";
 import TextField from "@material-ui/core/TextField";
-import logo from "./logo.png";
+import logo from "../../design/logo.png";
 
 class Login extends React.Component {
   state = {
